requirements.txt
fastapi
uvicorn
httpx
ollama
python-dotenv
pypdf
mcp